Update the code dealing with javascript added via ADD_JS_FILES.  This is the code currently in webwork, and preserves order.
Also updated the simple_format.pl to use bootstrap 5.  Instead of using
math4.css and math5.js a new js/apps/Problem/problem.css file is added
that has styles specifically for problem rendering.

The MathQuill editor and knowl javascript and styles are updated to work
both with ww3 and with the standalone renderer.

The template is also set to use the javascript in node_modules instead
of from a cdn.