Update the code that parses files added via ADD_JS_FILES to the current code in webwork2.  This preserves order of the requested files as well as honors the attributes that are passed.
I also noticed that the bootstrap javascript was not loaded in the
formats.  This is needed for scaffolds and results table popovers to
work.  The popovers need to be activated as well.  This is now done in
the math4.js file.  The pertinent styles were also added in the
math4.css file.

I removed the java_init.js file loaded in the templates.  That is
certainly not needed.