Remove the html id's from the problemMainForm and problem-content div in the ww3 format.  Those will cause id clashes if multiple problems are on the same page.  A unique form id will be added for each problem in ww3, and the problem_body id should never have existed anyway.
Update the ww_applet_support.js javascript for this.  The javascript now
adds the applet support submit handler to all forms. This means that the
submitAction for all applets will be called when any problem is
submitted.  This should work in general as usually the submit handler is
just transfering data from the applet to the hidden answer box.  It
won't hurt if that happens when it is not needed.  Also, if the problem
has the state reset button, then the javascript searches the forms on
the page to find the one that contains this particular reset button and
uses the correct preview button is used.