Add a setting via render_app.conf.dist to set static files from webwork2_files to be cached by the browser, and to expire a set amount of time in the future. This should reduce server load regarding static files which are used by multiple problems, but still makes it reasonable to roll out updated files in a reasonable manner.
At present, we only add the response header settings to support
browser caching for files under webwork2_files. However, we set expired
times for those whose URL contains a match to tmp/renderer as files
which were generated for a specific problem may be dangerous to cache,
in case they get updated in any manner on a subsequent call.