Update the renderer for the restructuring of PG.
The WeBWorK::PG module is now part of PG, and PG has its own environment
so the webwork2 WeBWorK::PG module (and its derivatives) have been
removed.

The WeBWorK::Constants, WeBWorK::Debug, and WeBWork::CourseEnvironment
modules and the webwork2 conf files are no longer needed and so are
removed.  The renderer uses the WeBWorK::PG::Environment module which is
part of PG.  Note that this adds one minor installation step for the
renderer.  The file conf/pg_config.yml must be copied to lib/PG/conf.
Usually that file will work as is, but in some cases administrators may
want to make changes to it.  For example, the externalPrograms are the
standard linux executables and in the standard locations.  On some
systems those may need to be changed.

Most of the WeBWorK::Utils module has been removed.  All that is left is
the two methods that the renderer uses.

In addition the renderer lib directory has been flattened.
lib/WeBWorK/lib/WeBWorK is now just lib/WeBWorK.  With this (among other
things) the WEBWORK_ROOT environment variable is not needed.  Note that
$WeBWorK::Constants::PG_DIRECTORY is not available.  $ENV{PG_ROOT} is
used instead in the renderer code.

Note that the unnecessary encoding and decoding of perl warnings has
been removed.

Also note that the references to drdrew42 have been changed to
openwebwork.