Remove the STATIC_EXPIRES option.
The option is set up to set an expiration for static assets served from
the `webwork2_files` route.  However, there is no such route anymore for
the renderer, and so the option currently does nothing.  This could be
adapted to work for the other files in public, but there is not point in
that.  The static assets used in production are either local assets that
are minimized and served with a hash in the file name, or are third
party assets and a version parameter is appended as a query parameter.
As such, when files change the browser will get the new versions.