Switch the static files to using a simplified `path_is_subdir` method from webwork2.
The problem with using `realpath` is that static assets can use symbolic
links.  In fact PG does this with static assets that reside in the same
directory as the problem file by default.  The `realpath` will not be a
subdirectory of the temporary directory in that case.  So `canonpath`
must be used instead.  It does not resolve symbolic links, or even
things like `../` in paths.  Of course the `path_is_subdir` checks those
things.